refactor(obstacle_stop_planner): reparate utils function
